+++ src/BootpayCommerceApi.php
@@ -36,7 +36,7 @@ class BootpayCommerceApi
 
     private static $postMethods = array('POST', 'PUT');
     private static $apiVersion = '2.5.0';
-    private static $sdkVersion = '2.6.1';
+    private static $sdkVersion = '2.6.2';
 
     public $user;
     public $userGroup;